I finished adding the shop, upgrades, and I also updated the scoring for the limbs
I may have used chatgpt for some of this but I believe it is important to use AI as tools to complete things faster. I used it to help me get some balanced values and equations in a way faster time while still taking the time to understand everything. (Also AI is NOT taking my job as it forgets what I ask like 3 messages later)